requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.7

# Authentication
django-allauth==0.57.0
dj-rest-auth==5.0.2
djangorestframework-simplejwt==5.3.0

# WebSockets & Real-time
channels==4.0.0
channels-redis==4.1.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# API Documentation
drf-yasg==1.21.7

# File uploads & Media
Pillow==10.0.1

# Environment variables
python-decouple==3.8

# Google API for Gemini
google-generativeai==0.3.1

# Additional utilities
django-extensions==3.2.3
django-debug-toolbar==4.2.0